Update encryption keys and modify secret configuration
This commit updates the encryption keys used in the secrets configuration, specifically for the `auth-key` and `mac` fields. Additionally, adds a new server named "ramen" in the `.sops.yaml` creation rules. Also, it removes `keyFile` and `generateKey` from the `age` module configuration in `default.nix`. Lastly, `lastModified` and `narHash` values in `flake.lock` file are updated to reflect these changes.